Move test in existing file
